chore(types): move global.d.ts to types/ and errors.ts to frontend utils; update imports and configs
