+++ frontend/src/scripts/charts.ts
@@ -1,7 +1,7 @@
 import Chart from 'chart.js/auto';
 import type { Chart as ChartJS } from 'chart.js'; // Импортируем тип ChartJS для аннотаций типов
 import { API_BASE_URL } from './api';
-import type { ErrorItem, Stats, PeriodStats } from './types/errors';
+import type { ErrorItem, Stats, PeriodStats } from './utils/errors';
 import { request } from './utils/request';
 import { qs, delegate } from './utils/dom';
 import { t, getLabel, onLangChange, setLang } from './utils/i18n';
