+++ frontend/src/scripts/header.ts
@@ -1,7 +1,7 @@
 import { ErrorApi } from './api';
 import { ErrorTable } from './table'; // Импорт типа конкретного менеджера графиков (ранее использовали ambient-интерфейс из global.d.ts)
 import { StatsManager } from './stats';
-import type { ErrorItem } from './types/errors';
+import type { ErrorItem } from './utils/errors';
 import type { ChartManagerType } from './charts';
 import { filterErrors } from './services/errorFilter';
 import { qs, createElement, delegate, translateNodes } from './utils/dom';
