+++ frontend/src/scripts/main.ts
@@ -7,7 +7,7 @@ import ChartManager from './charts';
 import type { ChartManagerType } from './charts'; // Импорт типа ChartManager чтобы приводить созданный инстанс к правильному типу (вместо ambient-описание в global.d.ts)
 import type { Aside } from './aside'; // Импорт типа Aside для приведения динамически загруженного синглтона
 import { ErrorTable } from './table';
-import type { ErrorItem, NewError } from './types/errors';
+import type { ErrorItem, NewError } from './utils/errors';
 import { onLangChange } from './utils/i18n';
 import { updateTestErrorButtonVisibility } from './utils/testErrorButton';
 import handleModuleLoadError from './utils/moduleLoad';
