+++ frontend/src/scripts/modal.ts
@@ -1,7 +1,7 @@
 import { el, setChildren } from 'redom';
 import { ErrorApi } from './api';
 import type { Mode } from './api';
-import type { ErrorItem } from './types/errors';
+import type { ErrorItem } from './utils/errors';
 // Импортируем тип `ErrorTable`, потому что вызываем `fetchErrors` на глобальном экземпляре и делаем локальное приведение `window.errorTableInstance` -> `ErrorTable`.
 import type { ErrorTable } from './table';
 import { qs, createElement, translateNodes, assertExists } from './utils/dom';
