+++ frontend/src/scripts/utils/errors.ts
