+++ frontend/tsconfig.json
@@ -9,7 +9,7 @@
     "checkJs": false,
     "noEmit": true
   },
-   "include": ["src/**/*"],
+  "include": ["src/**/*", "../types/**/*.d.ts"],
   // "include": ["src/**/*.ts", "src/**/*.tsx", "src/**/*.d.ts"],
   // "include": ["src/**/*", "src/scripts/types/**/*"],
   "exclude": [
