+++ tsconfig.base.json
@@ -4,6 +4,7 @@
     "module": "ES2022",
     "moduleResolution": "node",
     "esModuleInterop": true,
+    "typeRoots": ["./types", "./node_modules/@types"],
     "forceConsistentCasingInFileNames": true,
     "strict": true,
     "skipLibCheck": true,
